Force running the tests against polyfill implementation
Running the test suite in browsers that fully implement `classList` was a
moot point because all we were doing was exercising the native
implementation, which we already trust that it works (except in a few
browsers where we have to patch it). The actual code in the classList
implementation provided by this polyfill was never tested, unless you
were to load the test suite in an ancient browser.

This change instructs `script/test` to run the test suite a 2nd time by
pretending that the property name is now `testClassList` instead of
`classList`. Because browsers won't have the `testClassList` property
natively, this means that the implementation will actually get exercised
in tests, making sure that CI fails if someone breaks the implementation.

This is somewhat hackish, but it seems impossible to undefine or in any
way override the native `classList` property, so a new property name has
to be used.