Fix broken `classList.remove` test
The test used to first read the `el.classList` property (instantiating
the polyfilled DOMTokenList) and then setting `el.className`, but that
setting wouldn't get reflected on the already instantiated classList
object and the test failed when exercising the actual polyfill
implementation.

The test wouldn't fail with native `classList` implementation because
natively, a DOMTokenList is a live mirror of `className` property.